requirements.txt
flask
flask-cors
python-dotenv
requests
PyYAML
langchain
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
numpy
chromadb

